pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiogram-dialog>=2.3.1",
    "alembic>=1.16.1",
    "pydantic-settings>=2.9.1",
    "pyjson>=1.4.1",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
]
